refactor project page, externalize invitation column
